Moves Menu to events folder and rename it
